+++ src/vue/modules/create-asset-form/components/advanced-step-form.spec.js
@@ -26,7 +26,7 @@ describe('Advanced step form', () => {
     })
 
     const expectedResults = {
-      preissuedAssetSigner: ['required'],
+      preIssuanceAssetSigner: ['required'],
       initialPreissuedAmount: ['required', 'amountRange'],
     }
 
@@ -38,7 +38,7 @@ describe('Advanced step form', () => {
     }
 
     const fieldBindings = {
-      '[name=create-asset-preissued-asset-signer]': 'preissuedAssetSigner',
+      '[name=create-asset-pre-issuance-asset-signer]': 'preIssuanceAssetSigner',
       '[name=create-asset-initial-preissued-amount]': 'initialPreissuedAmount',
     }
 
@@ -83,7 +83,7 @@ describe('Advanced step form', () => {
       it('properly sets request prop fields to form property', () => {
         wrapper.setProps({
           request: {
-            preissuedAssetSigner: 'SIGNER_ID',
+            preIssuanceAssetSigner: 'SIGNER_ID',
             initialPreissuedAmount: '500.000000',
           },
         })
@@ -91,7 +91,7 @@ describe('Advanced step form', () => {
         wrapper.vm.populateForm()
 
         expect(wrapper.vm.form.isPreissuanceDisabled).to.be.false
-        expect(wrapper.vm.form.preissuedAssetSigner).to.equal('SIGNER_ID')
+        expect(wrapper.vm.form.preIssuanceAssetSigner).to.equal('SIGNER_ID')
         expect(wrapper.vm.form.initialPreissuedAmount).to.equal('500.000000')
       })
     })