+++ src/vue/modules/create-asset-form/components/advanced-step-form.vue
@@ -20,12 +20,12 @@
           <div class="advanced-step-form__pre-issued-asset-signer-wrp">
             <input-field
               white-autofill
-              v-model="form.preissuedAssetSigner"
-              @blur="touchField('form.preissuedAssetSigner')"
-              name="create-asset-preissued-asset-signer"
-              :label="'create-asset-form.preissued-signer-lbl' | globalize"
+              v-model="form.preIssuanceAssetSigner"
+              @blur="touchField('form.preIssuanceAssetSigner')"
+              name="create-asset-pre-issuance-asset-signer"
+              :label="'create-asset-form.pre-issuance-signer-lbl' | globalize"
               :error-message="getFieldErrorMessage(
-                'form.preissuedAssetSigner',
+                'form.preIssuanceAssetSigner',
               )"
               :disabled="isDisabled"
             />
@@ -34,11 +34,15 @@
               type="button"
               class="app__button-flat advanced-step-form__insert-account-id-btn"
               :disabled="isDisabled"
-              @click="form.preissuedAssetSigner = accountId"
+              @click="form.preIssuanceAssetSigner = publicKey"
             >
               {{ 'create-asset-form.use-my-account-id-btn' | globalize }}
             </button>
           </div>
+
+          <p class="advanced-step-form__pre-issuance-signer-disclaimer">
+            {{ 'create-asset-form.pre-issuance-signer-disclaimer' | globalize }}
+          </p>
         </div>
       </div>
 
@@ -125,14 +129,14 @@ export default {
   props: {
     request: { type: CreateAssetRequest, default: null },
     isDisabled: { type: Boolean, default: false },
-    accountId: { type: String, required: true },
+    publicKey: { type: String, required: true },
     maxIssuanceAmount: { type: String, default: '0' },
   },
 
   data: _ => ({
     form: {
       isPreissuanceDisabled: false,
-      preissuedAssetSigner: '',
+      preIssuanceAssetSigner: '',
       initialPreissuedAmount: '',
       terms: null,
     },
@@ -143,7 +147,7 @@ export default {
   validations () {
     return {
       form: {
-        preissuedAssetSigner: {
+        preIssuanceAssetSigner: {
           required: requiredUnless(function () {
             return this.form.isPreissuanceDisabled
           }),
@@ -170,13 +174,13 @@ export default {
   methods: {
     populateForm () {
       const isPreissuanceDisabled =
-        this.request.preissuedAssetSigner === config().NULL_ASSET_SIGNER
+        this.request.preIssuanceAssetSigner === config().NULL_ASSET_SIGNER
 
       this.form = {
         isPreissuanceDisabled: isPreissuanceDisabled,
-        preissuedAssetSigner: isPreissuanceDisabled
+        preIssuanceAssetSigner: isPreissuanceDisabled
           ? ''
-          : this.request.preissuedAssetSigner,
+          : this.request.preIssuanceAssetSigner,
         initialPreissuedAmount: isPreissuanceDisabled
           ? ''
           : this.request.initialPreissuedAmount,
@@ -223,5 +227,10 @@ export default {
 
 .advanced-step-form__insert-account-id-btn {
   margin-left: .4rem;
+}
+
+.advanced-step-form__pre-issuance-signer-disclaimer {
+  font-size: 1.4rem;
+  margin-top: 1rem;
 }
 </style>